# server/swagger.yaml
openapi: 3.0.3
info:
  title: Milk Dairy Management System API
  description: |
    A comprehensive MERN stack application for managing milk dairy operations with role-based access control, 
    subscription management, and real-time delivery tracking.
    
    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## User Roles
    - **ADMIN**: Full system access, user management, analytics
    - **DISTRIBUTOR**: Order management, delivery assignment, product management
    - **DELIVERY_BOY**: Delivery assignments, status updates
    - **CONSUMER**: Place orders, manage subscriptions, track deliveries
  version: 1.0.0
  contact:
    name: API Support
    email: support@milkdairy.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.milkdairy.com/api
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with role-based access
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              name: "John Doe"
              email: "john@example.com"
              password: "password123"
              phone: "9876543210"
              role: "CONSUMER"
              region: "North Delhi"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and get JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "john@example.com"
              password: "password123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid credentials or account deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      description: Retrieve the authenticated user's profile information
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/profile:
    put:
      tags:
        - Authentication
      summary: Update user profile
      description: Update the authenticated user's profile information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Address Management Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
        - phone
        - role
      properties:
        name:
          type: string
          minLength: 1
          description: User's full name
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          minLength: 6
          description: User's password (minimum 6 characters)
        phone:
          type: string
          pattern: '^[0-9]{10}$'
          description: User's phone number (10 digits)
        role:
          type: string
          enum: [ADMIN, DISTRIBUTOR, DELIVERY_BOY, CONSUMER]
          description: User's role in the system
        region:
          type: string
          description: Region (for distributors)
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          description: User's password
    AuthResponse:
      type: object
      properties:
        message:
          type: string
        token:
          type: string
          description: JWT authentication token
        user:
          $ref: '#/components/schemas/UserBasic'
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          description: User's full name
        phone:
          type: string
          pattern: '^[0-9]{10}$'
          description: User's phone number
        address:
          type: object
          description: User's address (deprecated - use addresses endpoint)

    # User Schemas
    User:
      type: object
      properties:
        id:
          type: string
          description: User ID
        name:
          type: string
          description: User's full name
        email:
          type: string
          format: email
          description: User's email address
        phone:
          type: string
          description: User's phone number
        role:
          type: string
          enum: [ADMIN, DISTRIBUTOR, DELIVERY_BOY, CONSUMER]
          description: User's role
        addresses:
          type: array
          items:
            type: string
          description: Array of address IDs
        isActive:
          type: boolean
          description: Whether the user account is active
        assignedDistributor:
          type: string
          description: ID of assigned distributor (for consumers)
        region:
          type: string
          description: Region (for distributors)
        profileImage:
          type: string
          description: URL to profile image
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    UserBasic:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        phone:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Error message
        error:
          type: string
          description: Detailed error information (development only)
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          description: Validation errors
...